Added missing fields
Added Missing category & urgency
