+++ server-java-mysql/src/main/java/com/example/groupproject/models/TicketRepository.java
@@ -2,6 +2,6 @@ package com.example.groupproject.models;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChatRepository extends JpaRepository<Chat, Long> {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
 
 }
